+++ releasenotes/notes/pipeline-hierarchy-b980efcf8ac6b122.yaml
@@ -1,0 +1,5 @@
+---
+enhancements:
+  - |
+    Create a class hierarchy for pipeline classes, and move the run logic into the child class.
+    Preparation work for introducing multiple run stratgegies.
